update render and cleanup engine

+++ module/src/com/vittach/jumpjack/ArrayIndex.java
@@ -1,9 +1,9 @@
 package com.vittach.jumpjack;
 
 class ArrayIndex {
-    int x, y, z;
+    float x, y, z;
 
-    ArrayIndex(int x, int y, int z) {
+    ArrayIndex(float x, float y, float z) {
         this.x = x;
         this.y = y;
         this.z = z;
